Switch from tensorflow to dl4j. Remove deprecated features. Add download request

+++ src/main/kotlin/org/jetbrains/research/groups/ml_methods/deepbugs/utils/Mapping.kt
@@ -3,5 +3,5 @@ package org.jetbrains.research.groups.ml_methods.deepbugs.utils
 import com.beust.klaxon.JsonObject
 
 class Mapping(private val json: JsonObject?) {
-    fun get(fieldName: String) = json?.array<Float>(fieldName)
+    fun get(fieldName: String) = json?.array<Double>(fieldName)
 }
